signaling_server/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
Flask-SocketIO==5.3.2
eventlet==0.33.2
gevent-websocket==0.10.1
python-dotenv==0.21.0